+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 6.0.4 (unreleased)
 ------------------
 
+- Add an option to limit the number of results shown.
+  [Gagaro]
+
 - Add a setting to limit the number of characters shown before and after.
   [Gagaro]
 
@@ -12,7 +15,6 @@ Changelog
 
 - Include toggleSelect for Plone 5.
   [Gagaro]
-
 
 
 6.0.3 (2016-02-15)